Merge upstream 21

+++ scripts/emailings/mailing-send.php
@@ -6,6 +6,7 @@
  * Copyright (C) 2005-2016 Regis Houssin <regis.houssin@inodbox.com>
  * Copyright (C) 2019 		Nicolas ZABOURI	<info@inovea-conseil.com>
  * Copyright (C) 2024       Frédéric France             <frederic.france@free.fr>
+ * Copyright (C) 2024		MDW							<mdeweerd@users.noreply.github.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -65,7 +66,13 @@ require_once $path."../../htdocs/master.inc.php";
 require_once DOL_DOCUMENT_ROOT.'/core/lib/functionscli.lib.php';
 require_once DOL_DOCUMENT_ROOT."/core/class/CMailFile.class.php";
 require_once DOL_DOCUMENT_ROOT."/comm/mailing/class/mailing.class.php";
-
+/**
+ * @var Conf $conf
+ * @var DoliDB $db
+ * @var HookManager $hookmanager
+ * @var Translate $langs
+ * @var User $user
+ */
 // Global variables
 $version = DOL_VERSION;
 $error = 0;
@@ -396,7 +403,7 @@ if ($resql) {
 								}
 
 								if (getDolGlobalInt('MAILING_DELAY')) {
-									usleep((float) getDolGlobalInt('MAILING_DELAY') * 1000000);
+									usleep((int) ((float) getDolGlobalInt('MAILING_DELAY') * 1000000));
 								}
 							}
 						} else {
